Show a loading screen when verifying auth status and fix server cronJobs

+++ server/config/cronJob.js
@@ -9,17 +9,19 @@ const updateTaskStatus = async () => {
 
     const tasks = await Task.find();
 
-    tasks.forEach(
-        async ({id, deadline_date, status}) => {
-        if (isAfter(currentDate, deadline_date) && status === 'pending') {
-            // change status of task to overdue
-            const data = { id, status: "overdue" };
-            const updatedTask = await Task.update(data);
-            // console.log(updatedTask);
-            // console.log("deadline date:", deadline_date);
-            // console.log("current date:", currentDate);
-        }
-    })
+    if (tasks) {
+        tasks.forEach(
+            async ({id, deadline_date, status}) => {
+            if (isAfter(currentDate, deadline_date) && status === 'pending') {
+                // change status of task to overdue
+                const data = { id, status: "overdue" };
+                const updatedTask = await Task.update(data);
+                // console.log(updatedTask);
+                // console.log("deadline date:", deadline_date);
+                // console.log("current date:", currentDate);
+            }
+        })
+    }
 }
 
 // Permanently deletes a task if the current date exceeds the deleted at date by 20 days.
@@ -28,20 +30,22 @@ const deleteTask = async () => {
 
     const tasks = await Task.find();
 
-    tasks.forEach(
-        async ({ id, deleted_at, archived }) =>  {
-            if (archived) {
-                const difference = differenceInDays(currentDate, deleted_at);
-
-                if (difference >= 30) {
-                    const deletedTask = await Task.deleteById(id);
-                    // console.log("deletedTask:", deletedTask);
-                    // console.log("deleted_at:", deleted_at);
+    if (tasks) {
+        tasks.forEach(
+            async ({ id, deleted_at, archived }) =>  {
+                if (archived) {
+                    const difference = differenceInDays(currentDate, deleted_at);
+    
+                    if (difference >= 30) {
+                        const deletedTask = await Task.deleteById(id);
+                        // console.log("deletedTask:", deletedTask);
+                        // console.log("deleted_at:", deleted_at);
+                    }
+                    // console.log("difference:", difference);
                 }
-                // console.log("difference:", difference);
             }
-        }
-    )
+        )
+    }
 }
 
 // For every 1 minute
